pyproject.toml
[project]
name = "stardew-valley-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Japoncio3k",email = "gustavo.kamihara@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0.0"
dependencies = [
    "langchain (>=1.0.8,<2.0.0)",
    "langchain[openai] (>=1.0.8,<2.0.0)",
    "langsmith (>=0.4.46,<0.5.0)",
    "unstructured (>=0.18.20,<0.19.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "qdrant-client (>=1.16.0,<2.0.0)",
    "fastembed (>=0.7.3,<0.8.0)",
    "sentence-transformers (>=3.0.0,<4.0.0)",
    "fastapi[standard] (>=0.115.0,<1.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.0.0,<4.3.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "uvicorn (>=0.42.0,<0.43.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = ">=0.9.0"
mypy = ">=1.0.0"
pytest = ">=8.0.0"
httpx = ">=0.27.0"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
